+++ src/main/java/com/example/nd_medicine/repository/AvailabilityRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AvailabilityRepository extends JpaRepository<Availability, Long> {
-    List<Availability> findAvailabilitiesByDoctorId(Long id);
+public interface AvailabilityRepository extends JpaRepository<Availability, Integer> {
+    List<Availability> findAvailabilitiesByDoctorId(Integer id);
 
     List<Availability> findByDoctorId(Doctor doctor);
