+++ src/main/java/com/example/nd_medicine/repository/SlotRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SlotRepository extends JpaRepository<Slot,Long> {
+public interface SlotRepository extends JpaRepository<Slot,Integer> {
 
     Slot findByDoctor(Doctor doctor);
